Move locker implementations to db-specific places
