+++ relstorage/adapters/postgresql/__init__.py
@@ -19,7 +19,7 @@ from relstorage.adapters.connmanager import AbstractConnectionManager
 from relstorage.adapters.dbiter import HistoryFreeDatabaseIterator
 from relstorage.adapters.dbiter import HistoryPreservingDatabaseIterator
 from relstorage.adapters.interfaces import IRelStorageAdapter
-from relstorage.adapters.locker import PostgreSQLLocker
+from .locker import PostgreSQLLocker
 from relstorage.adapters.mover import ObjectMover
 from relstorage.adapters.oidallocator import PostgreSQLOIDAllocator
 from relstorage.adapters.packundo import HistoryFreePackUndo
